// src/Types/formData.types.ts
interface IFormData {
  name: string;
  email: string;
  phoneNumber: string;
  address: string;
  schoolName: string;
  degree: string;
  eduStartDate: string;
  eduEndDate: string;
  company: string;
  position: string;
  workStartDate: string;
  workEndDate: string;
  location: string;
  description: string;
}
export type { IFormData };